% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_albedo}
\alias{df_albedo}
\title{Albedo lookup table}
\format{
A data frame with albedo values indexed by latitude and longitude
}
\usage{
df_albedo
}
\description{
Lookup table for Mars surface albedo based on latitude and longitude.
Based on NASA Technical Memorandum data.
}
\keyword{internal}
