% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_netflux_0p1_lookup_v1}
\alias{df_netflux_0p1_lookup_v1}
\title{Net flux lookup table (albedo 0.1, version 1)}
\format{
A data frame with net flux values indexed by zenith angle and optical depth
}
\usage{
df_netflux_0p1_lookup_v1
}
\description{
Lookup table for normalized net flux function with albedo fixed at 0.1.
From NASA TM-102299 (Appelbaum & Flood, 1989).
}
\keyword{internal}
