% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_netflux_k1_coeffs}
\alias{df_netflux_k1_coeffs}
\title{Net flux polynomial coefficients (k=1)}
\format{
A data frame with polynomial coefficients
}
\usage{
df_netflux_k1_coeffs
}
\description{
Polynomial coefficients for normalized net flux function calculation (albedo-dependent terms).
Based on Appelbaum & Flood analytical expression.
}
\keyword{internal}
