% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Normalized net flux function}
\usage{
f(z, tau, al = 0.1)
}
\arguments{
\item{z}{Sun zenith angle [deg]}

\item{tau}{Atmospheric optical depth (dimensionless)}

\item{al}{Surface albedo (dimensionless, ranges from 0.1 to 0.4)}
}
\value{
Normalized net flux (dimensionless)
}
\description{
Calculates the normalized net solar flux on the Martian surface accounting for multiple
wavelength and multiple scattering in the atmosphere. Based on Pollack's calculations
presented in Appelbaum & Flood (1990). Can use polynomial expression or lookup tables.
}
