% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_ob.R
\name{G_ob}
\alias{G_ob}
\title{Beam irradiance at top of Martian atmosphere}
\usage{
G_ob(Ls)
}
\arguments{
\item{Ls}{Areocentric longitude [deg]}
}
\value{
Beam irradiance at top of atmosphere [W/m²]
}
\description{
Calculates the solar beam irradiance at the top of the Martian atmosphere (before any
atmospheric effects) as a function of Mars' orbital position. Accounts for Mars' elliptical
orbit which causes seasonal variation in solar intensity. Implements Equation 4 from
Appelbaum & Flood (1990).
}
