% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_bh.R
\name{I_bh}
\alias{I_bh}
\title{Beam insolation on Mars horizontal surface over time period}
\usage{
I_bh(Ls, phi, tau, Ts_start, Ts_end)
}
\arguments{
\item{Ls}{Areocentric longitude [deg]}

\item{phi}{Planetary latitude [deg]}

\item{tau}{Atmospheric optical depth (dimensionless)}

\item{Ts_start}{Start of integration period [h]}

\item{Ts_end}{End of integration period [h]}
}
\value{
Beam insolation over specified time period [Wh/m²]
}
\description{
Calculates the direct beam solar energy received on a horizontal surface between specified
start and end times. Implements Equation 19 from Appelbaum & Flood (1990).
}
