% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polar_night.R
\name{is_polar_night}
\alias{is_polar_night}
\title{Check if location is experiencing polar night}
\usage{
is_polar_night(Ls, phi)
}
\arguments{
\item{Ls}{Areocentric longitude [deg]}

\item{phi}{Planetary latitude [deg]}
}
\value{
TRUE if experiencing polar night, FALSE otherwise
}
\description{
Determines whether a given location on Mars is experiencing polar night (24-hour darkness)
for the specified season. During polar night, the sun remains below the horizon continuously.
}
