% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars-package.R
\docType{package}
\name{mars-package}
\alias{mars}
\alias{mars-package}
\title{mars: Mars Solar Radiation}
\description{
A set of functions to calculate solar irradiance and insolation on Mars horizontal
and inclined surfaces. Based on NASA Technical Memorandum 102299, 103623, 105216,
106321, and 106700, i.e. the canonical Mars solar radiation papers.
}
\section{Package Information}{

\describe{
  \item{Version}{1.0.0}
  \item{License}{GPL-3}
  \item{Author}{Georges Labrèche <georges@tanagraspace.com> (ORCID: 0000-0002-2104-8789)}
  \item{Depends}{R (>= 2.10)}
}
}

\author{
\strong{Maintainer}: Georges Labrèche \email{georges@tanagraspace.com} (\href{https://orcid.org/0000-0002-2104-8789}{ORCID})

}
\keyword{internal}
