% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_angle.R
\name{optimal_angle}
\alias{optimal_angle}
\title{Optimal tilt angle for maximum daily insolation}
\usage{
optimal_angle(Ls, phi, unit = 1)
}
\arguments{
\item{Ls}{Areocentric longitude [deg]}

\item{phi}{Planetary latitude [deg]}

\item{unit}{Output unit: 1 for radians, 2 for degrees (default: 1)}
}
\value{
Optimal tilt angle [rad] or [deg] depending on unit parameter
}
\description{
Calculates the optimal surface tilt angle (beta) that maximizes daily solar energy collection
for a given location and season on Mars. The surface is assumed to face the equator.
Based on Equations 40 and 43 from Appelbaum (1993).
}
