% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise_for_horizontal_surface}
\alias{sunrise_for_horizontal_surface}
\title{Sunrise for horizontal surface}
\usage{
sunrise_for_horizontal_surface(phi, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunrise hour angle [rad]
}
\description{
Calculates sunrise hour angle for a horizontal surface.
Implements Equation 8 from Appelbaum (1993).
}
