% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise_for_inclined_surface}
\alias{sunrise_for_inclined_surface}
\title{Sunrise for inclined surface (general)}
\usage{
sunrise_for_inclined_surface(phi, beta, gamma_c, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{beta}{Surface tilt angle [rad]}

\item{gamma_c}{Surface azimuth angle [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunrise hour angle [rad]
}
\description{
Calculates sunrise hour angle for an inclined surface with any orientation.
Dispatches to specific orientation functions based on gamma_c value.
}
