% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_horizontal_surface}
\alias{sunset_for_horizontal_surface}
\title{Sunset for horizontal surface}
\usage{
sunset_for_horizontal_surface(phi, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunset hour angle [rad]
}
\description{
Calculates sunset hour angle for a horizontal surface.
Implements Equation 9 from Appelbaum (1993).
}
