% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_inclined_surface}
\alias{sunset_for_inclined_surface}
\title{Sunset for inclined surface (general)}
\usage{
sunset_for_inclined_surface(phi, beta, gamma_c, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{beta}{Surface tilt angle [rad]}

\item{gamma_c}{Surface azimuth angle [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunset hour angle [rad]
}
\description{
Calculates sunset hour angle for an inclined surface with any orientation.
Dispatches to specific orientation functions based on gamma_c value.
}
