% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_inclined_surface_oriented_east}
\alias{sunset_for_inclined_surface_oriented_east}
\title{Sunset for inclined surface oriented eastward}
\usage{
sunset_for_inclined_surface_oriented_east(phi, beta, gamma_c, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{beta}{Surface tilt angle [rad]}

\item{gamma_c}{Surface azimuth angle [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunset hour angle [rad], or NA if sun never sets on this surface
}
\description{
Calculates sunset hour angle for an inclined surface facing east.
Implements Equation 16 from Appelbaum (1993).
}
