% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_inclined_surface_oriented_equator}
\alias{sunset_for_inclined_surface_oriented_equator}
\title{Sunset for inclined surface oriented toward equator}
\usage{
sunset_for_inclined_surface_oriented_equator(phi, beta, delta)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{beta}{Surface tilt angle [rad]}

\item{delta}{Solar declination angle [rad]}
}
\value{
Sunset hour angle [rad]
}
\description{
Calculates sunset hour angle for an inclined surface facing the equator.
Implements Equation 31 from Appelbaum (1993).
}
