% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_d.R
\name{T_d}
\alias{T_d}
\title{Number of Mars daylight hours}
\usage{
T_d(Ls, phi)
}
\arguments{
\item{Ls}{Areocentric longitude [deg]}

\item{phi}{Planetary latitude [deg]}
}
\value{
Number of daylight hours [h]
}
\description{
Calculates the duration of daylight (time between sunrise and sunset) for a given
location and season on Mars. Implements Equation 10 from Appelbaum & Flood (1990).
}
