% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_for_inclined_surface.R
\name{x_for_inclined_surface}
\alias{x_for_inclined_surface}
\title{X parameter for inclined surface sunrise/sunset calculations}
\usage{
x_for_inclined_surface(phi, beta, gamma_c)
}
\arguments{
\item{phi}{Planetary latitude [rad]}

\item{beta}{Surface inclination angle [rad]}

\item{gamma_c}{Surface azimuth angle [rad]}
}
\value{
X parameter (dimensionless)
}
\description{
Calculates the X parameter used in determining sunrise and sunset times for inclined surfaces.
Implements Equations 19 and 25 from Appelbaum (1993) for inclined and vertical surfaces.
}
