% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albedo.R
\name{albedo}
\alias{albedo}
\title{The albedo function.}
\usage{
albedo(latitude, longitude, tau, coordinates_rounding = TRUE)
}
\arguments{
\item{latitude}{Planetary latitude [deg], from -90\u00b0 to +90\u00b0}

\item{longitude}{Planetary longitude [deg], from -180\u00b0 to 180\u00b0}

\item{tau}{Atmospheric optical depth (dimensionless)}

\item{coordinates_rounding}{Logical, whether to round coordinates to nearest grid point}
}
\value{
Surface albedo value (dimensionless, 0-1)
}
\description{
Calculate the albedo value given geographical location and tau factor.
Source: Appelbaum, Joseph & Landis, Geoffrey & Sherman, I. (1991).  Solar radiation on Mars — Update 1991.
}
